[Fix] - 캐릭터 상세페이지에서 새로고침시 무한 렌더링이 걸리는 버그 수정정
